Move same input different components pipeline
