+++ test/core/pipeline/features/pipeline_run.feature
@@ -22,6 +22,7 @@ Feature: Pipeline running
         | that has different combinations of branches that merge and do not merge |
         | that has two branches, one of which loops back |
         | that has a component with mutable input |
+        | that has a component with mutable output sent to multiple inputs |
 
     Scenario Outline: Running a bad Pipeline
         Given a pipeline <kind>
